Move General settings to `menus_settings_general.cpp`
